% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktaucenters_run.R
\docType{data}
\name{B1_DEFAULT}
\alias{B1_DEFAULT}
\title{ktaucenters_run}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
B1_DEFAULT
}
\arguments{
\item{data}{A matrix of size n x p.}

\item{K}{The number of clusters.}

\item{centers}{matrix of size K x p containing the K initial
centers, one at each matrix-row.}

\item{tolerance}{tolerance parameter used for the algorithm stopping
rule}

\item{max_iter}{a maximum number of iterations used for the
algorithm stopping rule}
}
\value{
A list including the estimated K centers and labels for the
    observations

\itemize{
\item{\code{centers}}{: matrix of size K
    x p, with the estimated K centers.}
\item{\code{cluster}}{:
    array of size n x 1 integers labels between 1 and K.}
\item{\code{tauPath}}{: sequence of tau scale values at each
    iterations.}
\item{\code{Wni}}{: numeric array of size n x 1
    indicating the weights associated to each observation.}
\item{\code{empty_cluster_flag}}{: a boolean value. True means
    that in some iteration there were clusters totally empty}
\item{\code{niter}}{: number of iterations until convergence
    is achived or maximum number of iteration is reached}
\item{\code{di}}{distance of each observation to its assigned
    cluster-center} }
}
\description{
Robust Clustering algorithm based on centers, a robust and
efficient version of K-Means.
}
\note{
Some times, if the initial centers are wrong, the algorithm
    converges to a non-optimal (local) solution.  To avoid that,
    the algorithm must be run several times. This task is carried
    out by \code{\link{ktaucenters}}
}
\examples{

# Generate Synthetic data (three cluster well separated)
Z=rnorm(600);
mues=rep(c(0,10,20),200)
data= matrix(Z+mues,ncol=2)

# Applying the algorithm
sal = ktaucenters_aux(
data, K=3, centers=data[sample(1:300,3), ],
tolerance=1e-3, max_iter=100)

#plot the results
plot(data,type='n')
points(data[sal$cluster==1,],col=1);
points(data[sal$cluster==2,],col=2);
points(data[sal$cluster==3,],col=3);

}
\references{
Gonzalez, J. D., Yohai, V. J., & Zamar, R. H. (2019).
    Robust Clustering Using Tau-Scales. arXiv preprint
    arXiv:1906.08198.
}
\seealso{
\code{\link{ktaucenters}}
}
\keyword{datasets}
