% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_tolerance}
\alias{max_tolerance}
\title{Estimates maximum tolerance between two given matrix row wise}
\usage{
max_tolerance(x, y)
}
\description{
Estimates maximum tolerance between two given matrix row wise
}
