% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_opt}
\alias{psi_opt}
\title{psi_opt function}
\usage{
psi_opt(t, c)
}
\arguments{
\item{t}{a numeric vector}

\item{c}{a tunning constant.}
}
\value{
psi(\code{t}/ \code{c})
}
\description{
An implementation of the derivative of quasi optimal rho function
}
\examples{
val <- psi_opt(t = 0.5, c = 1)

}
