% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_opt}
\alias{rho_opt}
\title{rho_opt function}
\usage{
rho_opt(t, c)
}
\arguments{
\item{t}{numeric vector.}

\item{c}{tunning constant.}
}
\value{
rho(\code{t}/\code{c})
}
\description{
An implementation of quasi optimal rho functions following reference [1]
}
\examples{
val <- rho_opt(t = 0.5, c = 1.0)
}
\references{
[1] Salibian-Barrera, M., Willems, G., & Zamar,
  R. (2008). The fast-tau estimator for regression. Journal of
  Computational and Graphical Statistics, 17(3), 659-682.
}
