% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derpsi_opt}
\alias{derpsi_opt}
\title{Implementation of the second derivative of the rho function}
\usage{
derpsi_opt(x, c)
}
\arguments{
\item{x}{a numeric vector}

\item{c}{a tunning constant.}
}
\value{
Numeric vector with with the second derivative of the quasi optimal rho
computation for each element of x.
}
\description{
Implementation of the second derivative of the rho function
}
