% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mscale}
\alias{mscale}
\title{mscale
the m scale of an univariate sample (see reference below)}
\usage{
mscale(u, b, c)
}
\arguments{
\item{u}{an univariate sample of size n.}

\item{b}{the desired break down point}

\item{c}{a tuning constant, if consistency to standard normal
  distribution is desired use
\code{\link{normal_consistency_constants}}}
}
\value{
the mscale value
}
\description{
mscale
the m scale of an univariate sample (see reference below)
}
