% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_opt}
\alias{psi_opt}
\title{Implementation of the derivative of quasi optimal rho function}
\usage{
psi_opt(x, c)
}
\arguments{
\item{x}{a numeric vector}

\item{c}{a tunning constant.}
}
\value{
Numeric vector with with the derivative of the quasi optimal rho computation
for each element of x.
}
\description{
Implementation of the derivative of quasi optimal rho function
}
