% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_opt}
\alias{rho_opt}
\title{Quasi optimal rho function following reference [1]}
\usage{
rho_opt(x, c)
}
\arguments{
\item{x}{numeric vector.}

\item{c}{tunning constant.}
}
\value{
Numeric vector with with quasi optimal rho computation for each element of
x.
}
\description{
Quasi optimal rho function following reference [1]
}
\references{
[1] Salibian-Barrera, M., Willems, G., & Zamar, R. (2008). The fast-tau
estimator for regression. Journal of Computational and GraphicalStatistics,
17(3), 659-682.
}
